+++ contracts/modules/STO/CappedSTOFactory.sol
@@ -13,9 +13,12 @@ contract CappedSTOFactory is ModuleFactory {
      * @param _polyAddress Address of the polytoken
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
-      ModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
-
+    ModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
+    {   
+        version = "1.0.0";
+        name = "CappedSTO";
+        title = "Capped STO";
+        description = "Use to collects the funds and once the cap is reached then investment will be no longer entertained";
     }
 
      /**
@@ -45,21 +48,35 @@ contract CappedSTOFactory is ModuleFactory {
      * @notice Get the name of the Module
      */
     function getName() public view returns(bytes32) {
-        return "CappedSTO";
+        return name;
     }
 
     /**
      * @notice Get the description of the Module
      */
     function getDescription() public view returns(string) {
-        return "Capped STO";
+        return description;
     }
 
     /**
      * @notice Get the title of the Module
      */
     function getTitle() public view returns(string) {
-        return "Capped STO";
+        return title;
+    }
+
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
+    }
+
+    /**
+     * @notice Get the setup cost of the module
+     */
+    function getSetupCost() external view returns (uint256) {
+        return setupCost;
     }
 
     /**