+++ contracts/modules/TransferManager/ManualApprovalTransferManagerFactory.sol
@@ -16,9 +16,12 @@ contract ManualApprovalTransferManagerFactory is ModuleFactory {
      * @param _subscriptionCost Subscription cost of the module
      */
     constructor (address _polyAddress, uint256 _setupCost, uint256 _usageCost, uint256 _subscriptionCost) public
-      ModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
-    {
-
+    ModuleFactory(_polyAddress, _setupCost, _usageCost, _subscriptionCost)
+    {   
+        version = "1.0.0";
+        name = "ManualApprovalTransferManager";
+        title = "Manual Approval Transfer Manager";
+        description = "Manage transfers using single approvals / blocking";
     }
 
      /**
@@ -44,21 +47,35 @@ contract ManualApprovalTransferManagerFactory is ModuleFactory {
      * @notice Get the name of the Module
      */
     function getName() public view returns(bytes32) {
-        return "ManualApprovalTransferManager";
+        return name;
     }
 
     /**
      * @notice Get the description of the Module
      */
     function getDescription() public view returns(string) {
-        return "Manage transfers using single approvals / blocking";
+        return description;
     }
 
     /**
      * @notice Get the title of the Module
      */
     function getTitle() public view returns(string) {
-        return "Manual Approval Transfer Manager";
+        return title;
+    }
+    
+    /**
+     * @notice Get the version of the Module
+     */
+    function getVersion() public view returns(string) {
+        return version;
+    }
+
+    /**
+     * @notice Get the setup cost of the module
+     */
+    function getSetupCost() external view returns (uint256) {
+        return setupCost;
     }
 
     /**